backend/spotify_app/serializers.py
from rest_framework import serializers

class SongSerializer(serializers.Serializer):
    title = serializers.CharField(max_length=255)
    artist = serializers.CharField(max_length=255)
    album = serializers.CharField(max_length=255)
    uri = serializers.URLField()