backend/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
python-dotenv>=0.19,<1.0
requests>=2.25,<3.0
django-cors-headers>=3.7,<4.0
django-environ>=0.9.0